Remove the scroll view paging extension
